+++ DataAccessLibrary/Models/UserWalletsModel.cs
@@ -9,7 +9,7 @@ namespace DataAccessLibrary.Models
     public class UserWalletsModel
     {
         public string UserId { get; set; }
-        public int PLN { get; set; }
+        public decimal PLN { get; set; }
         public int USD { get; set; }
         public int EUR { get; set; }
         public int CHF { get; set; }
